chore: оставил один файд utils. Перенес процедуры работы с git на верхний уровень
