+++ src/paqlang/functions/other_opers.py
@@ -3,7 +3,7 @@ import asyncio
 import aiofiles  # noqa
 
 from ..utils import aio_get_json, send_message, abs_file_path
-from .util_opers import coalesce, get_condition
+from ..utils import coalesce, get_condition
 
 from ..param import Param
